Sensors: Move objectPropertiesToJson() to sensor_utils
This commit move objectPropertiesToJson() to the sensor_utils class.
This is in preparation for building sensor excerpts using the same code
which builds a sensor. By moving the function into the sensor utility
class it should help with compile time by not needing to include all of
sensor code by the files implementing the excerpts.

Additional functions and definitions were moved into the sensor_utils
class as well due to dependencies from objectPropertiesToJson().

Tested:
 - Confirmed output unchanged for following URI:
   - /redfish/v1/Chassis/chassis/Sensors
   - Selection of: /redfish/v1/Chassis/chassis/Sensors/<sensorId>
   - /redfish/v1/Chassis/chassis/Sensors?\$expand=*
 - Redfish Validator passes
